bf16 true

+++ fit.py
@@ -69,7 +69,7 @@ def main(data_dir: Path, checkpoint_dir: Path):
         strategy="auto",
         max_epochs=epochs,
         deterministic=True,
-        precision="bf16-mixed",
+        precision="bf16-true",
         logger=WandbLogger(
             project="bistro",
             config=hparams,
